code/ai_pipeline/requirements.txt
clip_interrogator==0.6.0
ctransformers==0.2.27
diffusers==0.16.0.dev0
Flask==3.0.2
langchain==0.1.11
matplotlib==3.4.2
numpy==1.22.2
opencv_python==4.7.0.72
Pillow==10.2.0
Requests==2.31.0
torch==2.0.0
tqdm==4.66.1
